Merge Training in Colab/requirements.txt
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.24.3
google-generativeai==0.3.0
rank-bm25==0.2.2
